docs/tutorials/requirements.txt
jupyter
llnl-sina
kosh
matplotlib
maestrowf
merlin
seaborn
nbconvert
scipy
h5py
llnl-ibis
trata
scikit-learn
tensorflow
llnl-pydv
scipy
pyaml
pandas
jinja2
encorewf
numpy<2
